Diagnostics: Move types into a separate file.
This makes it easier to import things (e.g. the category enum) from places
without pulling in the diagnostics backend.
